requirements.txt
numpy>=1.21.0
scipy>=1.7.0
opencv-python>=4.5.0
librosa>=0.9.0
mediapipe>=0.8.0
pillow>=8.0
PyYAML>=5.4
ffmpeg-python>=0.2.1
playwright>=1.40.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
